refactor(chat-sessions): extract session write and title operations.
